add calibration of accelerometer and gyro data
- calibration original code: https://github.com/kriswiner/MPU9250
- accel data is now sent as a json float
- added marg_viz which shows the data

+++ src/sensor_reader/include/sensor_reader/mpu9250.hpp
@@ -19,9 +19,9 @@ class mpu9250 : public sensor
   public:
     struct data_t
     {
-        int16_t ax;  //!< Accelerometer X
-        int16_t ay;  //!< Accelerometer Y
-        int16_t az;  //!< Accelerometer Z
+        float ax;  //!< Accelerometer X
+        float ay;  //!< Accelerometer Y
+        float az;  //!< Accelerometer Z
 
         int16_t gx;  //!< Gyro X
         int16_t gy;  //!< Gyro Y
